requirements.txt
pandas
sqlalchemy
beautifulsoup4
streamlit
lxml
plotly.express